smarttech-streaming/smarttech-streaming/pyproject.toml
[project]
name = "smarttech-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "debugpy>=1.8.18",
    "delta-spark>=4.0.0",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.0",
    "kafka-python>=2.3.0",
    "pyspark>=4.0.1",
]
